Pulling data from Firebase function added
Database.service.ts function getBooks was modified - it now 'connect'
bookList observable to the result of data fetching from firebase db.
Data is saved in array of Books with added key to the parameters.

Booklist.component.ts calls getBooks function after with it's onInit
function, then subscribes to the bookList - actual list of books is
always in Booklist components and changes accordingly to the firebase
database.