Component showing book category created
Book-cat-container component has been defined. It shows one subcategory
in booklist container. It takes category name and books as input data.
It creates slider with buttons - if there's to many books to display
them on one page, user can use the arrows to navigate.

Booklist component calls for two instancees of the books (Hroror and
fanttasy) - it will be updated in the future.

Container size in styles.css has been changed - it's width is increased.