User read lists implemented
Component Read-list was defined - including html, css and ts code.
Components gets readlist, sorts it for two lists (readlist and already
read). Every position has button to mark as read/to read and to remove
from the list. Every position is a button to the book which this
position is describing.

New model was defined for single posittion of readlist.

User-read-lists service has been updated with new function -
modifyListPosition.